Update build.sh
Add SocksPolicy lines to permit only localhost and the private subnet where the workstation connects from are allowed to use the ports opened on tor. This is generally a good idea and tor will complain at startup if a port is open on anything other than localhost without a SocksPolicy line being set.