Possibly transproxy kernel leak
There is (was?) a bug in the Linux kernel whereby transproxied packets whose connections are unexpectedly closed may send unproxied RST packets out as uid 0 in some circumstances ( https://lists.torproject.org/pipermail/tor-talk/2014-March/032507.html ). This could potentially jump even a physically isolated proxy.

I added an iptables rule to explicitly drop outbound packets with an invalid state, which seems to be the cleanest way of preventing this behavior. I did not confirm the behavior is still present in newer kernels, but it seems prudent to include just in case.